pkg/utils: relicense to dual Apache2 and LGPLv3+
In a sequence of commits

fc8e4c5fa19d11afb4d9ebed8a94082b7d9459ae
31b83a82a22956349d0df178d5a0d212af237871
7c3cd4c1e0be107195558059c604ebf1dc765d53

Heketi was relicensed from Apache to the following model:

- the rest api client code is under dual Apache2 or LGPLv3+
- all other parts (server, cli, tests, ...) are under dual LGPLv3+ or GPLv2

Now an oversight/mistake was made in that parts of the pkg/utils are
used in the client and hence apache-licensed projects that compile
in heketi client, like kubernetes, have a license problem since they
pull GPL code in. See:

https://github.com/heketi/heketi/issues/1279
https://github.com/kubernetes/kubernetes/issues/35557
https://github.com/kubernetes/kubernetes/issues/70802

This patch fixes the oversight by relicensing the remaining
pieces of pkg/utils to the same dual Apache2 or LGPLv3+ license
after those parts that are only used in the server have been
moved out of pkg/utils.

Resolves: #1279

Signed-off-by: Michael Adam <obnox@redhat.com>
Signed-off-by: John Mulligan <jmulligan@redhat.com>