apps: add file for operations cleanup functionality
The operations_cleanup.go is the home of the types needed to
implement the core of the operations clean up logic. Clean up
is a pass over all the stale or failed pending operation entries
in the db and cleaning them up based on the cleaning methods
provided by each operation type.

The content of this file is an initial stub/POC that exists
to demonstrate the core of the clean up logic.

Signed-off-by: John Mulligan <jmulligan@redhat.com>