apps: skip over "id not found" load error from ops in the db
If the pending op is "damaged" and can't be loaded into a valid
operation then skip over it gracefully. This allows any other
working ops in the db to be cleaned. Currently, there's
no special handling for these damaged ops and thus they'll
remain in the db for manual cleanup.

Signed-off-by: John Mulligan <jmulligan@redhat.com>