container: use heketi-start.sh and heketi.json from standard location
The entry-point for containers is `heketi-start.sh` which automatically
gets installed as `/usr/share/heketi/container/heketi-start.sh` with
`make install`. There is no need to have a 2nd copy of the file in the
container image, just use the standard location directly.

`heketi.json` is the default configuration file, and can be installed as
`/etc/heketi/heketi.json` so that everything uses this version by
default.

Signed-off-by: Niels de Vos <ndevos@redhat.com>