remoteexec: use error output for error contents
Historically, heketi used to convert the stderr of commands to go error
type and discard the originating error. The remoteexec library however
goes to lengths to retain all the information about a command including
the error output (stderr) and originating error. But most of the
preexisting heketi code wants to check the error output not the
originating error.

This change allows treating a Result type as an error, looking for error
output if it exists and then falling back to the content of the
originating error. When the adapter functions FirstErrorIndexed,
FirstError, and AnyError were added the important ErrorOutput was
forgotten and the originating error was used. Now it returns the
result-as-error. This change fixes at least one regression where the
error response of "gluster volume delete" was being scraped.

Signed-off-by: John Mulligan <jmulligan@redhat.com>