apps: support exporting (dumping) db to stdout
This change allows the DbDump function to support the common cli
convention of treating a single dash "-" as a signal to write
to stdout rather than a file name. This allows us to run the
command `heketi db export --dbfile heketi.json --jsonfile -`
and read the json content on stdout.

Signed-off-by: John Mulligan <jmulligan@redhat.com>