systemd: provide a heketi.env file with possible variables
The environment variables that can be set to influence the behaviour of
Heketi are not well documented. This is an attempt to capture the
variables that can be changed for the benefit of the heketi.service
systemd unit.

The systemd unit already tries to read variables from
/etc/heketi/heketi.env, but this file is not provided anywhere. If users
have a need to change the environment, they will need to create the file
and not mistype the name of the variable. Providing a file with the
defaults makes it less error prone.

Signed-off-by: Niels de Vos <ndevos@redhat.com>