apps: add cleaner func for marking stale ops in running server
Add a function to the op cleaner type that will allow it to
find and mark stale operations that have not yet been so
marked. This is needed in cases where a server is
infrequently restarted but has failed to update a pending
op in the db after some time.

Signed-off-by: John Mulligan <jmulligan@redhat.com>