cmdexec: introduce a new config setting gluster_cli_timeout
This is in seconds.

The configured value can be overridden by setting the
environment variable HEKETI_GLUSTER_CLI_TIMEOUT.

The default of this variable is set to 10 minutes,
which is higher than gluster's default cli timeout
of 2 minutes, because in systems with many volumes,
some of the operations take longer than 2 minutes,
and cli timing out is interpreted as a failure by
heketi, while in the backend, gluster is still
performing the operation.

Signed-off-by: Michael Adam <obnox@redhat.com>