executors: add condition file support to inject executor
Sometimes it is convenient (or needed) to make changes to
an inject hook while the server is running. The most simple and
expedient way to get that done is to check for the
presence of "condition files". If the condition file is
specified but non-existent the hook is disabled. If the
condition file is specified and exists the hook is
enabled.

Signed-off-by: John Mulligan <jmulligan@redhat.com>