main: remove "db delete-pending-entries" command
Unfortunately, the delete-pending-entries command was always
buggy and could leave the db in a bad state. Instead of trying
to fix it more, we're just going to drop it entirely and rely
on the auto-cleanup code, future enhancement to the auto-cleanup
code, and good old fashioned db export/edit/import to
resolve future db & pending operation entry issues.

Signed-off-by: John Mulligan <jmulligan@redhat.com>