Implement sentinel errors as real constants
The errors created with errors.New() can't be used in constant
values.  Hence the sentinel errors could be overwritten by accident.
However, pure strings can be stored as constant values.  This change
replaces the sentinel errors with an error type that is isomorph to a
pure string and consequently defines them as constant values.

Signed-off-by: Sven Anderson <sven@redhat.com>