apps: prevent volume cleanup from corrupting the database
The volume cleanup code (function expungeVolumeWithOp) used to re-read
the volume entry from the DB instead of operating on the in-memory copy
from the operation. Hence the in-memory copy of the volume-entry of
the operation was still containig the list of brick-ids from the original
attempt even after the completion of expungeVolumeWithOp().

In the case the cleanup was called from the rollback of a volume create
operation, when the operation was retried, the next Build() process
would just continue using the in-memory volume entry and add new bricks
to the existing list instead of starting over from an empty list.  When
the volume entry was saved to disk at the end of Build(), it would
contain the new brick ids and the old brick ids. Those old brick ids did
not point to existing brick entries, because those had been cleaned  up
correctly in the rollback. Hence the database was rendered inconsistent.

This patch fixes this problem by having expungeVolumeWithOp()
act on the volume entry from the operation instead of re-loading
it from the database.

Pair-Programmed-With: Raghavendra Talur <rtalur@redhat.com>

Signed-off-by: Michael Adam <obnox@redhat.com>
Signed-off-by: Raghavendra Talur <rtalur@redhat.com>