apps: update in-memory operation objects with db contents in CleanDone
In order to add more isolation between the steps of the operations
the CleanDone functions operated not on the in-memory copy of the
volume or block volume entries, but on a copy freshly fetched from
the db by the ID. This was done in order to prevent the types of
problems we we seeing like in issue #1397 where data is populated
outside of the transaction it is needed in.

However, this has in turn caused a new problem in the create volume
operation which is the only operation with internal retries.
This operation keeps an in-memory volume entry object and manipulates
it during the build step to place bricks, etc. While isolating
the clean done delete prevents leakage it did not account for the
existing leakage of this volume entry object across multiple retry
attempts. This had the effect of leaving stale references to brick
ids in the volume object across multiple retry attempts.

This change resets the in-memory copy to the content in the db
(before final delete) which will reset the bricks list across
retry attempts.

The more correct, longer term, but too extensive for the moment fix
would be to stop referencing a in-memory mutable object for the steps
of the operation but rather take a immutable request object and
more cleanly isolate each step of the operation from each other.

Signed-off-by: John Mulligan <jmulligan@redhat.com>