logging: grab output of 'lsof /path/to/brick' in case unmounting fails
There are regular problems reported where Heketi fails to remove the
bricks from a Gluster volume. This is most often caused by an open
filedescriptor in one of the brick processes. Debugging this needs to be
done in a timely fashion, as processes may get restarted and any trace
of the kept open file is removed.

By logging the output of 'lsof /path/to/brick' when unmounting fails, we
should be able to find the filenames that the brick processes did not
close when the Gluster volume got stopped and deleted.

Note that the output of 'lsof' is only captured when Heketi is
configured when 'debug_umount_failures' is set to 'true' in the
heketi.json configuration file. The extra logging is enabled in the
shipped config file. These logs should not impact the normal runtime,
and only on a failure additional details are obtained. Failures to
unmount are rare, but if it happens the logs are crucial for analyzing.

In addition to the configuration option in the heketi.json file, the
environment variable HEKETI_DEBUG_UMOUNT_FAILURES can be set to "true"
or "false" to override the setting from the configuration file.

Signed-off-by: Niels de Vos <ndevos@redhat.com>