apps: streamline runOperationAfterBuild
This rewrites runOperationAfterBuild to one streamlined
(re)try loop. As a result, there is only one call to Rollback
left, and just one call to Exec. Due to this added consistency,
now, we call MarkFailedIfSupported for all cases when
Rollback fails, and we consistently always return the
original error if Rollback fails, which requires one
test case to be adapted (TestRunOperationExecRetryRollbackFail).
As the comment in this test case shows, it was originally
even intended to be like this, but silently changed in a
later commit (ce9a70113556e7f94305d5c3856ad57bc19487aa).

Signed-off-by: Michael Adam <obnox@redhat.com>