extras: introduce delete_unowned_bricks.sh script for brick cleanup
This script is useful for cleaning of bricks when they are not in use
by any Gluster volume and heketi does not have any reference to them.
This is possible when the heketi database has been scrubbed of brick
entries to make the db consistent.

The script must be run on each node/pod of Gluster as it cleans up only
bricks local to the node. The script is run in two phases. In the first
phase(scan), a list is made of unused LVs on the node and files
containing the list are created per VG. In the second phase(delete),
the script is invoked with those files as input.

ip-prefix argument is required to determine bricks local to the node.

fstab-path argument is required and is usually "/var/lib/heketi/fstab"
in a pod and "/etc/fstab" if Gluster is not containerized.

NOTE: make sure heketi is not running while the script is run.

Before running the script in delete mode, check the files that it generated have valid IDs.

Pair-Programmed-With: John Mulligan <jmulligan@redhat.com>
Pair-Programmed-With: Raghavendra Talur <rtalur@redhat.com>

Signed-off-by: John Mulligan <jmulligan@redhat.com>
Signed-off-by: Raghavendra Talur <rtalur@redhat.com>