executors: handle bhv does not exist errors in block delete
In the case that we request a delete of a block volume but the
block hosting volume never existed the error response from gluster
block is a little different than the usual. In this case we
get "Volume X does not exist" which heketi was not treating as
a does-not-exist type error when processing gluster-block delete
response. This change handles that.

Signed-off-by: John Mulligan <jmulligan@redhat.com>