cli: print distribute & replica counts in volume info
The heketi-cli output was not informative with regards to the gluster
volume distribute count. This change clarifies that these values are
counts and displays the distribute count for all types along with the
specific counts, such as replica count, for the volume's durability type.

Peer-programmed-with: Michael Adam <obnox@redhat.com>
Signed-off-by: John Mulligan <jmulligan@redhat.com>