tests: temporarily disable goimports
It appears that in quite a few of our (totally valid) files the
goimports tool gets confused and removes valid imports.
The problem is similar to that mentioned in
  https://github.com/golang/go/issues/29557
but I am not entirely sure that's the exact issue.
So until all our files are goimports safe or goimports behavior
is improved I'm disabling the check script.

FWIW this script was never run in travis ci as it appears goimports
tool was not added to our travis instances and was not
immediately caught when the script was added.

Signed-off-by: John Mulligan <jmulligan@redhat.com>