testutils: stop trying to delete devices in parallel
Trying to delete devices in parallel is slowing the functional
tests down. This happens because its looping linearly over nodes
(typically 4) and then working in parallel across all that node's
devices (up to 8). Not only is this slightly strange point to
parallelize on, it can actually cause the operations throttle to
trigger and delay the test teardown.

Signed-off-by: John Mulligan <jmulligan@redhat.com>