client: move response polling delay to central function
Previously, every "api function" that needed to poll for a response
would pass a hard-coded timeout to the polling function. Thus, every
function was choosing its own delay. Most functions used 1 second but
some used 0.25 seconds. No obvious reason was given for these values.

This change moves the choice of the response polling delay out of the
individual api functions and into a common function and makes the delay
configurable at the time the client is set up by the ClientOptions
struct.

Signed-off-by: John Mulligan <jmulligan@redhat.com>