container: add ability to archive db at container start up
Add code to the container start-up script that uses the new
HEKETI_DB_ARCHIVE_PATH and HEKETI_DB_ARCHIVE_COUNT environment variables
to keep a limited set of archival copies of the heketi db as it was
before the server process started. If HEKETI_DB_ARCHIVE_PATH is set to
"/dev/null" or "-" the path will be unset and the code block is skipped.
Otherwise, a compressed copy of the db will be made, keeping up to
HEKETI_DB_ARCHIVE_COUNT copies of the db.
Only valid databases (currently determined by checking if db is
serializeable) are archived. If the current archive hashes to an
existing hash the newest archive is discarded and the original kept.

Signed-off-by: John Mulligan <jmulligan@redhat.com>