apps: use "last takes precedence" when extracting heketi opts
When extracting a volume option meant for heketi from the volume options
of a volume, use the more common "last item takes precedence" rule. This
will cause heketi to match what the options meant for gluster do and
matches with common expectations.

This change also creates a volume options to map function such that the
functions that extract/parse a value do not have to handle the common
aspects of searching through the options list to find keys and handle
the precedence individually.

Signed-off-by: John Mulligan <jmulligan@redhat.com>