client: add function for getting default client options
Add a function that is meant to get the same default client options as
the internal "new client" functions do. This is meant as light
future-proofing for client packages that want to get the same defaults
as most other client api libraries do but tweak the values of a few
fields.

Signed-off-by: John Mulligan <jmulligan@redhat.com>