executors: add function to "upgrade" an old handle
If the device entry was created before these recent changes it
will not persist the PV UUID in the db (or the alternate paths list).
Because the initial device name may not be the current device name
try to use the vg id (which ought to be persistent) to map back
to a PV UUID (and thus a persistent path).

Signed-off-by: John Mulligan <jmulligan@redhat.com>