apps: fix issue where old BHVs could not be cleaned up
Older versions of heketi did not deduct the size of a new block volume
from the new BHV in a compound BHV+BV create. During the implementation
of auto-cleanup the only BHV tested were "new" ones with the space
properly deducted. Running heketi with some older pending ops in the db
showed heketi was not cleaning these old ops. This change corrects the
auto-cleanup code to ignore updating the used sizes of the BVH when the
BHV is already pending (and thus to be deleted) during cleanup.

Fixes #1572

Signed-off-by: John Mulligan <jmulligan@redhat.com>