functional tests: add test cases for removing devices w/o metadata
These tests dump the db json and then strip the entries of the new
device ident metadata. Then the devices are deleted to ensure that
the devices can still be managed with only the "old" metadata.
The 2nd test case also uses "alias" symlinks to test that if the
symlink is removed/changed during the lifetime of the device in
the heketi db it can still be removed.

Signed-off-by: John Mulligan <jmulligan@redhat.com>