functional tests: check that "old" pending BHV+BV can be cleaned
This test mutates a current db with a BHV+BV pending create to look
similar enough to the old contents of the block hosting volumes info of
an old heketi version that didn't update the used size of the BHV in
build. With the previous fix to the cleanup code this test passes, but
without this test will fail.

Signed-off-by: John Mulligan <jmulligan@redhat.com>