remote exec: define a richer command type
Previously, all commands were strings and were passed to the executor in
slice based batches. This change enriches the Cmd type that was
previously stubbed in by converting to a more customizable struct that
can be added to over time.
Additionally, this puts heketi on the path to one day passing commands
as vectors (one string per argument) rather than one giant string.
This patch includes all of the touch points needed for the type
conversion.

Signed-off-by: John Mulligan <jmulligan@redhat.com>