heketi-cli: allow user supplied templates for topology info
It can frequently be useful to scrape output from topology info but the
output format, while generally useful as a human readable format is not
simple to parse. Allow the use of a go template [1] to be supplied on
the command line to custom format the api response within the topology
info command. This is similar in spirit to the kubectl -o go-template=xxx option.

1: https://golang.org/pkg/text/template/

Signed-off-by: John Mulligan <jmulligan@redhat.com>