apps: gracefully handle attempts to clean same op at once
Previously heketi's operation tracker used an "assertion" to ensure an
operation id was being inserted into the tracker only once. This worked
fine for APIs such as delete, create, etc which always create a new
operation (in general). However, operations cleanup examines the db for
(stale) operations and works on what was existing.  Thus two
simultaneous requests for operations cleanup could lead to the cleaner
requesting tracking (or throttling of) an operation already in the
tracker, leading to a panic.

This change alters the insert function to return an error when the id is
already being tracked so that the caller(s) can do something more
intelligent with the condition as needed.

Signed-off-by: John Mulligan <jmulligan@redhat.com>