testutils: add test "Helper" function to server control utility
Add test "Helper" to wrapper interface and to various test control
utility functions. This will cause tests that use "testing.T"
as the interface's type to be "ignored" so that the failing line
in the test file itself is printed.
(See https://golang.org/pkg/testing/#T.Helper for more info)

Signed-off-by: John Mulligan <jmulligan@redhat.com>