executors: add new config params for kubeexec around timeout handling
Now that kube exec backend supports timeout features allow them to be
configured by the kubeexec executor. We will enable the timeout command
prefix but allow the user to disable it. We will default to 64
concurrent kube connections but allow the user to disable the threshold
with a negative value or specify a custom value.

Signed-off-by: John Mulligan <jmulligan@redhat.com>