apps: remove old device registration feature
The device register feature would create special db entries with keys
based on the name (path) of the device. Now that heketi tracks device
meta data explicitly and this metadata can be used to avoid overwriting
devices in use, this feature doesn't protect against much. Additionally,
it interferes with the case where a new device is added that is now
using the same path as a previously added device.
Remove this approach to checking for previous device use and rely
entirely on the new executor-error-handling approach.

Signed-off-by: John Mulligan <jmulligan@redhat.com>