apps: only permit passing of destroy data flag when safe
Previously, the device destroy feature did not check if the device
was in use by heketi only that the name (path) was known. In the case
of devices that do not have persistent names this could lead to
destroying the data of a device that heketi was already managing.
Avoid this by checking the state from the error type the executor
returns.

Signed-off-by: John Mulligan <jmulligan@redhat.com>