extra tools: correctly handle cli "input" errors
At some point this code was refactored and code that used to have the
argparse parser object in scope was moved to new functions. This made
error-condition checks on the cli values post-parsing invalid.
This change fixes it by creating a new exception that the main function
with the parser object can then handle.

Signed-off-by: John Mulligan <jmulligan@redhat.com>