functional tests: refactor shell methods for testenv setup
Switch the functional test common library to use a new method for
setting up the test environment. Now we expect the callers to provide
env vars:
  HEKETI_TEST_ENVIRONMENT_UP
  HEKETI_TEST_ENVIRONMENT_DOWN
as the primary entry and exit points for the test environment.
By default this uses the new common vagrant env dir and the scripts:
  up.sh
  down.sh

Using an alternate or custom env is as simple as providing these two
env vars with executable commands that take no arguments.

Signed-off-by: John Mulligan <jmulligan@redhat.com>