apps: add a configuration option to force the app to open the db read-only
In order to change the app setup process as little as possible but to
avoid any risks with opening the db read-write we add a new field to the
app config to force the app to open the db read only. This will allow us
to use the entire app object (db + executor layers) in special purpose
recovery commands where we don't want to change the content of the db.

Signed-off-by: John Mulligan <jmulligan@redhat.com>