main: add an offline update-dbvol command
The "heketi offline update-dbvol" command is designed to change the
gluster volume options of heketidbstorage. If the volume is not
configured, the so called "level 0" then it will apply the new defaults
for heketidbstorage volumes (called level 1). Otherwise it will
make no other changes. The name of the volume to update can be forced for
testing purposes.

Signed-off-by: John Mulligan <jmulligan@redhat.com>