apps: don't check that an expanding volume is pending
The internal db consistency checker was incorrectly looking for the
volume being expanded to be marked pending. This is not the right
behavior as a pending item is considered incomplete/unusable and a
volume being expanded is not incomplete. Change the check to only
verify that the volume id in question exists not that it is pending.

Signed-off-by: John Mulligan <jmulligan@redhat.com>