apps: fix incorrect ID being used for expand vol reclaim map
The reclaim map maps device IDs to booleans to indicate that space was
reclaimed on the device by the removal of a brick. The volume expand
operation clean up code was incorrectly looking up in the map with the
brick id and thus, even in the basic case, the space from these bricks
was not being reclaimed on the device.

Note that this condition would only occur if the operation failed due
to a failure of the lvm/fs/gluster commands which is less common than
having the operation fail (early) when it is rejected by heketi for
low-space, too many bricks, etc.

Signed-off-by: John Mulligan <jmulligan@redhat.com>