fix inconsistent db after (failing) volume expand operation (#1673)
* apps: don't check that an expanding volume is pending

The internal db consistency checker was incorrectly looking for the
volume being expanded to be marked pending. This is not the right
behavior as a pending item is considered incomplete/unusable and a
volume being expanded is not incomplete. Change the check to only
verify that the volume id in question exists not that it is pending.

Signed-off-by: John Mulligan <jmulligan@redhat.com>

* apps: fix incorrect ID being used for expand vol reclaim map

The reclaim map maps device IDs to booleans to indicate that space was
reclaimed on the device by the removal of a brick. The volume expand
operation clean up code was incorrectly looking up in the map with the
brick id and thus, even in the basic case, the space from these bricks
was not being reclaimed on the device.

Note that this condition would only occur if the operation failed due
to a failure of the lvm/fs/gluster commands which is less common than
having the operation fail (early) when it is rejected by heketi for
low-space, too many bricks, etc.

Signed-off-by: John Mulligan <jmulligan@redhat.com>

* functional tests: check db is consistent after each cleanup

Add checks to the offline cleanup test cases to ensure the db is
consistent after each case is executed. Currently it is only used in the
tests that were being checked for a related issue but this could
certainly be duplicated for other test suites.

Signed-off-by: John Mulligan <jmulligan@redhat.com>