apps: create a new operation for brick eviction
Brick eviction is the act of "kicking out" a specified brick from a
volume and automatically replacing it. This operation aims to be robust
when the operation fails either before or after the brick is actually
replaced in glusterd's data store. This is why the cleanup paths are
split into to major paths. One handles the case where the new brick is
present in gluster and the other where the old brick is still present in
gluster.

Signed-off-by: John Mulligan <jmulligan@redhat.com>