apps: rework remove device so it uses evict brick child operations
Replace the old ad-hoc sequence of exec calls and db modifications made
by the device remove operation with iteration of calls into the newer
brick evict operation. This makes remove device a parent operation and
the brick evict operations children. The only interesting this about
this is the intermediate operation type created to "glue" the db changes
for the parent and child together in one transaction.

Modulo bugs this should be more reliable than the old approach because
brick evict's clean is smarter and properly handle failures or stale
operations. All the parent has to do it let the child handle the clean
up of the child operation that failed and then clean its own db state
up. Note that this means the operation is cleaned, but the operation is
not _resumed_. Any operation that is failed or incomplete will only
clean the current child and the device will remain in the offline state
and may still have one ore more bricks on it. But at this point it
should be safe to request removal of the device onace again.

Signed-off-by: John Mulligan <jmulligan@redhat.com>