apps: add parent and child pending operation change types
These "meta" change types allow us to link operations together in
parent-child relationships. Parent-child operations are operations
that run one operation as part of another. This will be used to update
the device remove operation so that it becomes a parent of iterated
brick evict operations, where all the bricks on the device are to
be evicted.

Signed-off-by: John Mulligan <jmulligan@redhat.com>