extras: update container start script with ability to update dbvol
Recently, we added a tool to heketi that sets the heketidbstorage volume
gluster settings to something more appropriate for hosting db files.
Currently that tool must be run by hand. This change adds the ability to
run the tool automatically during the start-up of the container.
To enable the tool set the environment var HEKETI_TRY_UPDATE_DB_VOL
to "yes".

Signed-off-by: John Mulligan <jmulligan@redhat.com>