cmdexec: prevent a crash by checking error return in diagnostic code
The code that checks to see if a brick is stuck from unmounting has a
bug in that it does not check the error returned by the command that
runs lsof. This was probably lazy refactoring on my part. Move the bulk
of the diagnostic section into a helper function and only try to use the
command result when there was no connection error.

Signed-off-by: John Mulligan <jmulligan@redhat.com>