travis-ci: set travis:travis as owner/group for pip cache
While setting up the test environment on Travis CI, pip is run as root.
This seems to create a .cache/pip directory in the travis users home
directory. Subsequently running pip as user travis fails:

    WARNING: Building wheel for d2to1 failed: [Errno 13] Permission denied: '/home/travis/.cache/pip/wheels/4d'

The pip the option --no-cache-dir does not seem to be sufficient.
Possibly pip is run at a different time, and the cache directory already
exists. Running 'chown' on the directory so that the next pip execution
as user travis as permission to store contents in the cache.

Signed-off-by: Niels de Vos <ndevos@redhat.com>