apps: chain the errors found when setting device state
The previous code had the odd behavior of returning a new error when the
no replacement error was returned, but that is only if the called code
returns exactly the error value expected. And, as it happens, the
existing does not. This is super confusing when writing tests so instead
of trying to keep everything straight, chain the errors so that
we can at least assert on the error text if needed.

Signed-off-by: John Mulligan <jmulligan@redhat.com>