apps: do not cache device id that's only used in one place
The cached device id value was only referenced (practically) in
one place, and that place was already inside a db transaction.
This caching has a bug in that the cached value is not updated in
the load-clean-clean-done path and therefore it is simpler to
just remove it entirely and use the device id from the db.

Signed-off-by: John Mulligan <jmulligan@redhat.com>