apps: have brick evict distinguish between unknown and brick-not-replaced
If brick eviction fails to run the checks against gluster in the clean
up path, its better to be able to tell in the later code if we ever got
the status or got a positive result for brick replace complete or not.
A boolean can't help us here so this switches to an enum where the "zero
value" is status-unknown.

Signed-off-by: John Mulligan <jmulligan@redhat.com>