functional tests: add test case for brick evict lvremove error
This test case forces the rollback to fail and leaves the pending
operation in the db. Then we explicitly ask it to be cleaned up.
Without recent fixes this exposes an error in the clean+clean-done
path because the brick allocation code is not run from an operation
loaded from the db. This passes with recent fixes in place.

Signed-off-by: John Mulligan <jmulligan@redhat.com>