go-client: add DialContext option to client
This adds a configurable DialContext function to the client API, which
will be invoked to create HTTP/HTTPS connections.

Allowing a custom DialContext captures a wide variety of use cases and
behaviors by delegating the responsibility to the DialContext function.
Another benefit is that it is a minimal change in go-client itself.

For example, users can set dial timeouts, share connections, or
tunnel/proxy connections using custom logic.

To support golang 1.10 through 1.12, this manually implements a clone
func for the default http transport.

Signed-off-by: Jonathan Basseri <misterikkit@google.com>