executors: add default functions for inject exec mock functions
The error injection executor's internal mock executor lacked functions
to return the NotSupportedError sentinel. Thus the error injection, or
any of the child executors would not get executed for Block volume
expand or block volume info as long as the inject executor was
in use. Fix this by supplying inject executor appropriate defaults.

Signed-off-by: John Mulligan <jmulligan@redhat.com>