executors: add a default error value for mock executor
The mock executor treated every executor call as a "silent success" by
default. This ran contrary to the way the error injection executor
was making use of the mock executor. Add a default error value to the
mock executor, defaulting to nil, that is returned by all the default
executor functions.

Signed-off-by: John Mulligan <jmulligan@redhat.com>