cli: avoid treating characters in db dump as format strings
Previously, the cli code for printing a db dump gathered from the server
was passing the string as the first string argument to a print function.
This print will then use any characters in the text from the server,
like %-signs, as format strings and potentially mangle the output. My
lazy fix is to just use a single "%s" format string and move the json
data from the server to a secondary argument.

Signed-off-by: John Mulligan <jmulligan@redhat.com>