apps: work around issue evicting & destroying old bricks on dead nodes
If a brick destroy calls fails during brick evict and the node is
down, do not treat it as a fatal error. If the node or device replace
is being done for a permanently down node then the destroy call will
never succeed. Use the ability of heketi to probe the node to decide
if if the error is fatal (node responsive) or not (node unresponsive).

There's one flaw with this: if the node is temporarily down. In that
case we could be orphaning bricks on said node, and when it comes up
that node would be out of sync with heketi. A future direction might
be to add a new node state to indicate "offline-and-never-coming-back"
that indicate that errors in this code path are truly safe to ignore.

Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit e0a5ea4ac7f2032a4cc839327fa93c03c8b9905a)