apps: try and use a working node where possible during brick eviction
There were a couple additional locations in the brick evict code paths
where we assumed that the node hosting the brick to be evicted was
online and able to accept gluster commands. Since this is not the case
when the node is dead, we need to make this a bit more robust. This
patch tries to identify all those locations and use a common helper
function to handle those cases.

Signed-off-by: John Mulligan <jmulligan@redhat.com>