client: allow consumers of heketi api to customize the http client
This change allows consumers of the heketi api to fully customize the
http client. By default, heketi continues behaving as it has. However,
the SetClientFunc call can be used to pass a function to the client that
will be used to get a client every time one is needed. The client
function actually requires a new interface, one that matches all of
heketi's needs of http client (currently only Do), allowing for extra
customization by returning something other than a true http.Client.

With these changes consumers of the heketi api should not be limited
by the options that heketi provides by default as these low-level
elements in the client stack can now be tweaked freely.

The HeketiHttpClient function is exported so that one can get a client
just as heketi would by default, but further customize that, if needed
rather than requiring the consumer code to create the client wholesale.

Signed-off-by: John Mulligan <jmulligan@redhat.com>