docs: remove inline python and ruby examples
Thanks to issue #1853 we determined that the in-line Python example was
incorrect (out of date). Since we have a python lib that is actually
tested (most of the time), we prefer that get used instead.

Similarly, the ruby example could be out of date and no one on the
project would no, because the maintenance team doesn't know ruby. Plus,
it clearly shells out to curl... and I sure hope that's not the
canonical way to do http in ruby.

Signed-off-by: John Mulligan <jmulligan@redhat.com>